Move getTablesWithReplicationModes to config
