+++ src/config/index.js
@@ -1,3 +1,4 @@
+const toPairs = require('lodash/toPairs');
 const extractConfigurationFromEnvironment = require('./extract-configuration-from-environment');
 
 const configuration = extractConfigurationFromEnvironment();
@@ -13,8 +14,22 @@ const repeatableJobOptions = {
   repeat: { cron: configuration.SCHEDULE, tz: parisTimezone },
 };
 
+const REPLICATION_MODE = {
+  INCREMENTAL: 'incremental',
+  TO_EXCLUDE: 'none',
+};
+
+function getTablesWithReplicationModes(configuration, modes = []) {
+  const tablePairs = toPairs(configuration.BACKUP_MODE);
+  return tablePairs
+    .filter(([_, mode]) => modes.includes(mode))
+    .map(([tableName, _]) => tableName);
+}
+
 module.exports = {
   configuration,
   jobOptions,
   repeatableJobOptions,
+  REPLICATION_MODE,
+  getTablesWithReplicationModes,
 };
